refactor: replace macro call with function call
Macros are often used to reduce the boilerplate code to write. In case
of the kernel call the actual code isn't that much different from the
macro call.

A reader of the source code might suspect something complex going on in
the macro, although it's really straight forward. Hence I decided to
just replace the macro call with the actual code.

This also eases future uses of the call as type errors won't be hidden
within a macro, but will just be regular errors.